move internal-helpers.ts
